day019_mnist-sketch/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.9
